backend/requirements.txt
Flask==3.1.1
flask_cors==6.0.1
pytest==8.4.1
psycopg2-binary==2.9.10